Fix: mock a11y api request for court events glossary (#140)
* fix: mock a11y api request for court events glossary

* use mock data for a11y tests

* fix: refactor names and increase code coverage

* fix: refactor names
