+++ src/main/service/statusDescriptionService.ts
@@ -1,12 +1,12 @@
-import {SearchDescriptionRequests} from '../resources/requests/searchDescriptionRequests';
-import {CourtService} from '../service/courtService';
+import {StatusDescriptionRequests} from '../resources/requests/statusDescriptionRequests';
+import {CourtService} from './courtService';
 
-const searchDescriptionRequests = new SearchDescriptionRequests();
+const statusDescriptionRequests = new StatusDescriptionRequests();
 
 export class StatusDescriptionService {
 
   public async generateStatusDescriptionObject(): Promise<object> {
-    const statusDescriptionList: Array<any>= await searchDescriptionRequests.getStatusDescriptionList();
+    const statusDescriptionList: Array<any>= await statusDescriptionRequests.getStatusDescriptionList();
     const alphabetOptions = CourtService.generateAlphabetObject();
 
     //Then loop through each status, and add it to the list
