+++ src/test/unit/service/statusDescriptionService.test.ts
@@ -3,16 +3,16 @@ import sinon from 'sinon';
 import { expect } from 'chai';
 import fs from 'fs';
 import path from 'path';
-import {SearchDescriptionRequests} from '../../../main/resources/requests/searchDescriptionRequests';
+import {StatusDescriptionRequests} from '../../../main/resources/requests/statusDescriptionRequests';
 
 const statusDescriptionService = new StatusDescriptionService();
 
-const searchDescriptionRequests = SearchDescriptionRequests.prototype;
+const statusDescriptionRequests = StatusDescriptionRequests.prototype;
 
 const rawData = fs.readFileSync(path.resolve(__dirname, '../mocks/StatusDescription.json'), 'utf-8');
 const statusDescriptionData = JSON.parse(rawData);
 
-const stub = sinon.stub(searchDescriptionRequests, 'getStatusDescriptionList').returns(statusDescriptionData);
+const stub = sinon.stub(statusDescriptionRequests, 'getStatusDescriptionList').returns(statusDescriptionData);
 stub.withArgs().returns(statusDescriptionData);
 
 const validStatusDescriptionKeysCount = 26;
